texture: Remove miniquad context from public API
This is very breaking change: all free-standing textures functions got removed. But now we do not have miniquad in public API and Textures API is consistent with Image API.
