+++ examples/platformer.rs
@@ -17,7 +17,7 @@ struct Platform {
 #[macroquad::main("Platformer")]
 async fn main() {
     let tileset = load_texture("examples/tileset.png").await.unwrap();
-    set_texture_filter(tileset, FilterMode::Nearest);
+    tileset.set_filter(FilterMode::Nearest);
 
     let tiled_map_json = load_string("examples/map.json").await.unwrap();
     let tiled_map = tiled::load_map(&tiled_map_json, &[("tileset.png", tileset)], &[]).unwrap();
