+++ examples/post_processing.rs
@@ -3,7 +3,8 @@ use macroquad::prelude::*;
 #[macroquad::main("Post processing")]
 async fn main() {
     let render_target = render_target(320, 150);
-    set_texture_filter(render_target.texture, FilterMode::Nearest);
+    render_target.texture.set_filter(FilterMode::Nearest);
+
     let material =
         load_material(CRT_VERTEX_SHADER, CRT_FRAGMENT_SHADER, Default::default()).unwrap();
